requirements.txt
streamlit
pandas
numpy
openai
requests
transformers
plotly
streamlit-lottie
pinecone-client
python-dotenv
